properly handle generics in all contexts
