+++ macros/src/types/newtype.rs
@@ -4,7 +4,7 @@ use syn::{FieldsUnnamed, Generics, Result};
 use crate::{
     attr::{FieldAttr, Inflection},
     deps::Dependencies,
-    types::generics::format_type,
+    types::generics::{format_generics, format_type},
     DerivedTS,
 };
 
@@ -48,8 +48,9 @@ pub(crate) fn newtype(
         None => format_type(inner_ty, &mut dependencies, generics),
     };
 
+    let generic_args = format_generics(generics).unwrap_or_default();
     Ok(DerivedTS {
-        decl: quote!(format!("type {} = {};", #name, #inline_def)),
+        decl: quote!(format!("type {}{} = {};", #name, #generic_args, #inline_def)),
         inline: inline_def,
         inline_flattened: None,
         name: name.to_owned(),
